商品接口

+++ src/main/java/com/bupt/recycle/entity/Category.java
@@ -19,7 +19,7 @@ public class Category {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long category_id;
+    private int category_id;
 
     private String category_name;
